move HeroGroupLabel to separate component
