+++ frontend-next-migration/src/entities/Hero/ui/HeroContainerTempByLeo/HeroContainer.module.scss
@@ -143,30 +143,6 @@
   max-height: 38%;
 }
 
-.heroGroupLabel{
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-  flex-direction: row;
-
-  &Title{
-    background-repeat: no-repeat;
-    background-position: center center;
-    background-attachment: scroll;
-    background-size: cover;
-    padding: 4vw;
-    text-shadow: 1px 1px 2px black;
-
-    & span {
-      font-size: calc(12px + 0.8vw);
-      display: inline-block;
-      transform: translateX(calc(12px + 0.2vw));
-      hyphens: manual;
-    }
-  }
-}
-
 .heroName {
   text-align: center;
   margin-bottom: 2rem;
