+++ frontend-next-migration/src/entities/Hero/ui/HeroGroupLabel/HeroGroupLabel.module.scss
@@ -1,0 +1,23 @@
+.heroGroupLabel{
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    text-align: center;
+    flex-direction: row;
+  
+    &Title{
+      background-repeat: no-repeat;
+      background-position: center center;
+      background-attachment: scroll;
+      background-size: cover;
+      padding: 4vw;
+      text-shadow: 1px 1px 2px black;
+  
+      & span {
+        font-size: calc(12px + 0.8vw);
+        display: inline-block;
+        transform: translateX(calc(12px + 0.2vw));
+        hyphens: manual;
+      }
+    }
+  }
